Renamed `build_ebook_v2.py` -> `build_ebook.py`
